+++ src/ng2-pagination.ts
@@ -3,6 +3,7 @@ import {CommonModule} from '@angular/common';
 import {PaginatePipe} from './paginate.pipe';
 import {PaginationService} from './pagination.service';
 import {PaginationControlsComponent} from './pagination-controls.component';
+import {PaginationControlsDirective} from './pagination-controls.directive';
 
 export {PaginationService, IPaginationInstance} from './pagination.service';
 export {PaginationControlsComponent} from './pagination-controls.component';
@@ -12,9 +13,10 @@ export {PaginatePipe} from './paginate.pipe';
     imports: [CommonModule],
     declarations: [
         PaginatePipe,
-        PaginationControlsComponent
+        PaginationControlsComponent,
+        PaginationControlsDirective
     ],
     providers: [PaginationService],
-    exports: [PaginatePipe, PaginationControlsComponent]
+    exports: [PaginatePipe, PaginationControlsComponent, PaginationControlsDirective]
 })
 export class Ng2PaginationModule { }